pyproject.toml
[tool.poetry]
name = "gpt-terminal-chat"
version = "0.1.0"
description = ""
authors = ["alejoboga-ai <alejo@zelta.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.354"
python-dotenv = "^1.0.0"
openai = "<=1.0.0"
chromadb = "^0.4.22"
tiktoken = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
